+++ bin/store/src/controllers/common_controller.rs
@@ -704,7 +704,7 @@ pub async fn process_and_get_record_by_id(
 pub async fn perform_upsert(
     table_name: &str,
     conflict_columns: Vec<String>,
-    mut data: serde_json::Value,
+    data: serde_json::Value,
     pluck_fields: Option<Vec<String>>,
     auth: &Auth,
 ) -> Result<ApiResponse, ApiError> {