+++ bin/store/src/providers/find/validations.rs
@@ -1,20 +1,18 @@
 // use serde::{Serialize, Deserialize};
 // use serde_json::Value;
-use crate::structs::structs::{
-    ApiResponse, GetByFilter,
-};
+use crate::structs::structs::{ApiResponse, GetByFilter};
 
 // #[derive(Serialize, Deserialize)]
 pub struct Validation<'a, 'b> {
     request_body: &'a GetByFilter,
-    table: &'b String
+    table: &'b String,
 }
 
 impl<'a, 'b> Validation<'a, 'b> {
     pub fn new(request_body: &'a GetByFilter, table: &'b String) -> Self {
         Self {
             request_body,
-            table
+            table,
         }
     }
     // Made this a static method since it doesn't need self
@@ -26,37 +24,40 @@ impl<'a, 'b> Validation<'a, 'b> {
     //         Vec::new()
     //     }
     // }
-    
+
     pub fn exec(&self) -> ApiResponse {
         // let mut keys = Self::get_keys_from_object(&self.request_body);
         // keys.append(&mut vec!["table".to_string()]);
         let required_keys = vec!["table", "pluck", "advance_filters:group_advance_filters"];
-        let mut response =  ApiResponse {
+        let mut response = ApiResponse {
             success: true,
-            message: format!("Successfully validated request body with keys: {}", required_keys.join(", ")),
+            message: format!(
+                "Successfully validated request body with keys: {}",
+                required_keys.join(", ")
+            ),
             count: required_keys.len() as i32,
-            data: vec![]
+            data: vec![],
         };
         // Iterate through keys and validate each property that are required before proceeding
         for key in &required_keys {
             match *key {
                 "table" => {
-                   response = self.validate_table();
-                   if !response.success {
-                     return response;
-                   }
+                    response = self.validate_table();
+                    if !response.success {
+                        return response;
+                    }
                 }
                 "pluck" => {
-                   response = self.validate_pluck();
-                   if !response.success {
-                     return response;
-                   }
+                    response = self.validate_pluck();
+                    if !response.success {
+                        return response;
+                    }
                 }
                 "advance_filters:group_advance_filters" => {
                     response = self.validate_conflicting_filters();
                     if !response.success {
-                     return response;
-                   }
+                        return response;
+                    }
                 }
                 // Add more validation cases as needed
                 _ => {
@@ -64,8 +65,8 @@ impl<'a, 'b> Validation<'a, 'b> {
                 }
             }
         }
-        
-       response
+
+        response
     }
 
     pub fn validate_table(&self) -> ApiResponse {
@@ -74,38 +75,40 @@ impl<'a, 'b> Validation<'a, 'b> {
                 success: false,
                 message: "table is required".to_string(),
                 count: 0,
-                data: vec![]
+                data: vec![],
             };
         }
-        
+
         ApiResponse {
             success: true,
             message: "Successfully validated table field".to_string(),
             count: 0,
-            data: vec![]
+            data: vec![],
         }
     }
 
-     pub fn validate_pluck(&self) -> ApiResponse {
+    pub fn validate_pluck(&self) -> ApiResponse {
         if self.request_body.pluck.is_empty() {
             return ApiResponse {
                 success: false,
                 message: "pluck is required".to_string(),
                 count: 0,
-                data: vec![]
+                data: vec![],
             };
         }
-        
+
         ApiResponse {
             success: true,
             message: "Successfully validated pluck field".to_string(),
             count: 0,
-            data: vec![]
+            data: vec![],
         }
     }
 
     pub fn validate_conflicting_filters(&self) -> ApiResponse {
-        if !self.request_body.advance_filters.is_empty() && !self.request_body.group_advance_filters.is_empty() {
+        if !self.request_body.advance_filters.is_empty()
+            && !self.request_body.group_advance_filters.is_empty()
+        {
             return ApiResponse {
                 success: false,
                 message: "Both advance_filters and group_advance_filters cannot be provided at the same time".to_string(),
@@ -118,7 +121,7 @@ impl<'a, 'b> Validation<'a, 'b> {
             success: true,
             message: "Successfully validated conflicting properties".to_string(),
             count: 0,
-            data: vec![]
+            data: vec![],
         }
     }
 }